+++ src/inthost.rs
@@ -10,6 +10,7 @@ use crate::bus::MessageBus;
 use crate::BindingsList;
 use crate::{authz, errors, Actor, Authorizer, NativeCapability, RouteKey};
 use errors::ErrorKind;
+use provider_archive::ProviderArchive;
 use std::str::FromStr;
 use std::{
     collections::HashMap,
@@ -454,9 +455,10 @@ pub(crate) fn fetch_oci_bytes(img: &str) -> Result<Vec<u8>> {
     let mut c = oci_distribution::Client::new(cfg);
 
     let img = oci_distribution::Reference::from_str(img).map_err(|e| {
-        crate::errors::new(crate::errors::ErrorKind::MiscHost(
-            "Could not parse OCI distribution reference".to_string(),
-        ))
+        crate::errors::new(crate::errors::ErrorKind::MiscHost(format!(
+            "Failed to parse OCI distribution reference: {}",
+            e
+        )))
     })?;
     let auth = if let Ok(u) = std::env::var(OCI_VAR_USER) {
         if let Ok(p) = std::env::var(OCI_VAR_PASSWORD) {
@@ -483,6 +485,12 @@ pub(crate) fn fetch_oci_bytes(img: &str) -> Result<Vec<u8>> {
             )))
         }
     }
+}
+
+pub(crate) fn fetch_provider_archive(img: &str) -> Result<ProviderArchive> {
+    let bytes = fetch_oci_bytes(img)?;
+    ProviderArchive::try_load(&bytes)
+        .map_err(|e| format!("Failed to load provider archive: {}", e).into())
 }
 
 fn invocation_from_callback(
@@ -610,8 +618,55 @@ pub(crate) fn detect_core_host_labels() -> HashMap<String, String> {
         CORELABEL_OSFAMILY.to_string(),
         std::env::consts::FAMILY.to_string(),
     );
-    trace!("Intrinsic host labels: {:?}", hm);
+    info!("Detected Intrinsic host labels. hostcore.arch = {}, hostcore.os = {}, hostcore.family = {}",
+          std::env::consts::ARCH,
+          std::env::consts::OS,
+          std::env::consts::FAMILY,
+    );
     hm
+}
+
+pub(crate) fn fetch_actor(actor_id: &str) -> Result<crate::actor::Actor> {
+    let vec = crate::inthost::fetch_oci_bytes(actor_id)?;
+
+    crate::actor::Actor::from_slice(&vec)
+}
+
+pub(crate) fn fetch_provider(
+    provider_ref: &str,
+    binding_name: &str,
+    labels: Arc<RwLock<HashMap<String, String>>>,
+) -> Result<(
+    crate::capability::NativeCapability,
+    Claims<wascap::jwt::CapabilityProvider>,
+)> {
+    use std::fs::File;
+    use std::io::Write;
+
+    let par = crate::inthost::fetch_provider_archive(provider_ref)?;
+    let lock = labels.read().unwrap();
+    let target = format!("{}-{}", lock[CORELABEL_ARCH], lock[CORELABEL_OS]);
+    let v = par.target_bytes(&target);
+    if let Some(v) = v {
+        let path = std::env::temp_dir();
+        let path = path.join(target);
+        {
+            let mut tf = File::create(&path)?;
+            tf.write_all(&v)?;
+        }
+        let nc = NativeCapability::from_file(path, Some(binding_name.to_string()))?;
+        if let Some(c) = par.claims() {
+            Ok((nc, c))
+        } else {
+            Err(format!(
+                "No embedded claims found in provider archive for {}",
+                provider_ref
+            )
+            .into())
+        }
+    } else {
+        Err(format!("No binary found in provider archive for {}", target).into())
+    }
 }
 
 #[cfg(test)]